+++ data/models/HAHM_asymmetric_UFO/CT_couplings.py
@@ -1,0 +1,11 @@
+# This file was automatically created by FeynRules 2.3.49
+# Mathematica version: 12.0.0 for Linux x86 (64-bit) (April 7, 2019)
+# Date: Sat 14 Sep 2024 14:34:15
+
+
+from object_library import all_couplings, Coupling
+
+from function_library import complexconjugate, re, im, csc, sec, acsc, asec, cot
+
+
+