+++ data/models/HAHM_asymmetric_UFO/coupling_orders.py
@@ -1,0 +1,24 @@
+# This file was automatically created by FeynRules 2.3.49
+# Mathematica version: 12.0.0 for Linux x86 (64-bit) (April 7, 2019)
+# Date: Sat 14 Sep 2024 14:34:15
+
+
+from object_library import all_orders, CouplingOrder
+
+
+NP = CouplingOrder(name = 'NP',
+                   expansion_order = 99,
+                   hierarchy = 1)
+
+HIG = CouplingOrder(name = 'HIG',
+                    expansion_order = 99,
+                    hierarchy = 1)
+
+QCD = CouplingOrder(name = 'QCD',
+                    expansion_order = 99,
+                    hierarchy = 1)
+
+QED = CouplingOrder(name = 'QED',
+                    expansion_order = 99,
+                    hierarchy = 2)
+